Use symlinks for all dotfiles

+++ link.sh
@@ -3,17 +3,9 @@
 for filename in .*; do
   if [ $filename != ".git" ] && [ $filename != ".gitmodules" ]  && [ $filename != "." ] && [ $filename != ".." ]
   then
-    if [ -d $filename ]
+    if [ ! -d ~/$filename ] && [ ! -f ~/$filename ]
     then
-      if [ ! -d ~/$filename ]
-      then
-        ln -s ~/.dotfiles/$filename ~/$filename
-      fi
-    else
-      if [ ! -f ~/$filename ]
-      then
-        ln $filename ..
-      fi
+      ln -s ~/.dotfiles/$filename ~/$filename
     fi
   fi
 done
